# Conflicts:
#	baro-cloud/gateway/src/main/resources/application.yml

+++ baro-support/src/main/java/com/barofarm/support/experience/domain/Reservation.java
@@ -13,7 +13,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.UUID;
 import lombok.AccessLevel;
@@ -64,7 +63,7 @@ public class Reservation extends BaseEntity {
     private Integer headCount;
 
     @Column(name = "total_price", nullable = false)
-    private BigInteger totalPrice;
+    private Long totalPrice;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
@@ -77,7 +76,7 @@ public class Reservation extends BaseEntity {
         LocalDate reservedDate,
         String reservedTimeSlot,
         Integer headCount,
-        BigInteger totalPrice,
+        Long totalPrice,
         ReservationStatus status
     ) {
         this.reservationId = reservationId;
@@ -102,7 +101,7 @@ public class Reservation extends BaseEntity {
             LocalDate reservedDate,
             String reservedTimeSlot,
             Integer headCount,
-            BigInteger totalPrice
+            Long totalPrice
     ) {
         if (reservedDate != null) {
             this.reservedDate = reservedDate;
